# backend/app/schemas/user.py
from __future__ import annotations

from datetime import datetime
from typing import Optional

from pydantic import BaseModel


class TelegramAuth(BaseModel):
    """Схема для входа/регистрации через Telegram."""
    telegram_id: int
    username: Optional[str] = None
    first_name: Optional[str] = None
    last_name: Optional[str] = None


class UserSettings(BaseModel):
    currency: Optional[str] = "RUB"
    language: Optional[str] = "ru"


class UserResponse(BaseModel):
    id: int
    telegram_id: int
    username: Optional[str] = None
    first_name: Optional[str] = None
    last_name: Optional[str] = None
    settings: Optional[UserSettings] = None
    created_at: datetime

    model_config = {"from_attributes": True}


class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"