# docs/FinMate-openapi-schema.yaml
openapi: 3.0.3
info:
  title: FinSight API
  description: API финансового помощника с автоматической категоризацией и учётом транзакций
  version: 1.0.0
  contact:
    name: Команда FinSight
    email: dev@finsight.example.com

servers:
  - url: http://localhost:8000/api/v1
    description: Локальный сервер разработки

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-токен, полученный через /auth/telegram

  schemas:
    # --- Аутентификация через Telegram ---
    TelegramAuth:
      type: object
      required: [telegram_id]
      properties:
        telegram_id:
          type: integer
          description: ID пользователя в Telegram
        username:
          type: string
          description: Telegram username (без @)
        first_name:
          type: string
        last_name:
          type: string

    UserSettings:
      type: object
      properties:
        currency:
          type: string
          example: RUB
        language:
          type: string
          example: ru

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        telegram_id:
          type: integer
        username:
          type: string
          nullable: true
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        settings:
          $ref: '#/components/schemas/UserSettings'
        created_at:
          type: string
          format: date-time

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer

    # --- Счета ---
    AccountCreate:
      type: object
      required: [name, type, balance]
      properties:
        name:
          type: string
          maxLength: 100
        type:
          type: string
          enum: [cash, bank, credit, investment, savings]
        balance:
          type: number
          format: double
          example: 15000.50
        currency:
          type: string
          default: RUB
        credit_limit:
          type: number
          format: double
          nullable: true

    AccountUpdate:
      type: object
      properties:
        name:
          type: string
        balance:
          type: number
          format: double
        is_archived:
          type: boolean
        credit_limit:
          type: number
          nullable: true

    AccountResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        type:
          type: string
        balance:
          type: number
        currency:
          type: string
        is_archived:
          type: boolean
        credit_limit:
          type: number
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Категории ---
    CategoryCreate:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          maxLength: 100
        type:
          type: string
          enum: [income, expense]
        parent_id:
          type: integer
          nullable: true
        icon:
          type: string
        color:
          type: string

    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        parent_id:
          type: integer
          nullable: true

    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        type:
          type: string
        parent_id:
          type: integer
          nullable: true
        icon:
          type: string
        color:
          type: string
        is_system:
          type: boolean

    # --- Транзакции ---
    TransactionCreate:
      type: object
      required: [account_id, type, amount, date]
      properties:
        account_id:
          type: integer
        category_id:
          type: integer
          nullable: true
          description: Если не указана, категория подбирается ML-сервисом
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
          format: double
          minimum: 0.01
        description:
          type: string
          maxLength: 500
        date:
          type: string
          format: date
        is_recurring:
          type: boolean
          default: false
        recurrence_rule:
          type: string
          description: RRULE для повторяющихся транзакций

    TransactionUpdate:
      type: object
      properties:
        category_id:
          type: integer
          nullable: true
        amount:
          type: number
          format: double
        description:
          type: string
        date:
          type: string
          format: date

    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        account_id:
          type: integer
        category_id:
          type: integer
          nullable: true
        type:
          type: string
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        notes:
          type: string
        date:
          type: string
          format: date
        is_recurring:
          type: boolean
        recurrence_rule:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string

  parameters:
    accountIdParam:
      name: account_id
      in: path
      required: true
      schema:
        type: integer
      description: ID счёта
    categoryIdParam:
      name: category_id
      in: path
      required: true
      schema:
        type: integer
      description: ID категории
    transactionIdParam:
      name: transaction_id
      in: path
      required: true
      schema:
        type: integer
      description: ID транзакции

# --------------------------------------------------------------------------
paths:
  # ========================= Аутентификация =========================
  /auth/telegram:
    post:
      tags: [Auth]
      summary: Вход/регистрация через Telegram
      description: >
        Если пользователь с таким telegram_id не найден — создаётся новый.
        Если найден — обновляются данные профиля и выдаётся JWT-токен.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramAuth'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Ошибка валидации

  # ========================= Пользователи =========================
  /users/me:
    get:
      tags: [Users]
      summary: Получить профиль текущего пользователя
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Не авторизован

  # ========================= Счета =================================
  /accounts:
    get:
      tags: [Accounts]
      summary: Список счетов текущего пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: is_archived
          in: query
          schema:
            type: boolean
          description: Фильтр архивных/активных счетов
      responses:
        '200':
          description: Массив счетов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
        '401':
          description: Не авторизован

    post:
      tags: [Accounts]
      summary: Создать новый счёт
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreate'
      responses:
        '201':
          description: Счёт создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Неверные данные

  /accounts/{account_id}:
    get:
      tags: [Accounts]
      summary: Получить счёт по ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/accountIdParam'
      responses:
        '200':
          description: Данные счёта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Счёт не найден

    put:
      tags: [Accounts]
      summary: Обновить счёт
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/accountIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: Счёт обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'

    delete:
      tags: [Accounts]
      summary: Удалить счёт
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/accountIdParam'
      responses:
        '204':
          description: Счёт удалён
        '404':
          description: Счёт не найден

  # ========================= Категории ============================
  /categories:
    get:
      tags: [Categories]
      summary: Список категорий пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
          description: Фильтр по типу
      responses:
        '200':
          description: Массив категорий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '401':
          description: Не авторизован

    post:
      tags: [Categories]
      summary: Создать пользовательскую категорию
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Категория создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: Ошибка валидации

  /categories/{category_id}:
    put:
      tags: [Categories]
      summary: Обновить категорию
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/categoryIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        '200':
          description: Категория обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '403':
          description: Нельзя изменить системную категорию
        '404':
          description: Категория не найдена

    delete:
      tags: [Categories]
      summary: Удалить категорию
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/categoryIdParam'
      responses:
        '204':
          description: Категория удалена
        '403':
          description: Нельзя удалить системную категорию
        '404':
          description: Категория не найдена

  # ========================= Транзакции ===========================
  /transactions:
    get:
      tags: [Transactions]
      summary: Список транзакций с фильтрацией
      security:
        - bearerAuth: []
      parameters:
        - name: account_id
          in: query
          schema:
            type: integer
        - name: category_id
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список транзакций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
        '401':
          description: Не авторизован

    post:
      tags: [Transactions]
      summary: Создать транзакцию (автоматически обновит баланс счёта)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '201':
          description: Транзакция создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Ошибка валидации или недостаточно средств
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /transactions/{transaction_id}:
    put:
      tags: [Transactions]
      summary: Обновить транзакцию (корректирует баланс при изменении суммы)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/transactionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdate'
      responses:
        '200':
          description: Транзакция обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Транзакция не найдена

    delete:
      tags: [Transactions]
      summary: Удалить транзакцию (возвращает сумму на счёт)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/transactionIdParam'
      responses:
        '204':
          description: Транзакция удалена
        '404':
          description: Транзакция не найдена